+++ CHANGELOG.md
@@ -4,6 +4,17 @@ Changelog
 Development version
 -------------------
 
+Version 0.12 (October xx, 2022)
+-------------------------------
+
+The highlight of this release is the support for Shapely 2.0. This makes it possible to
+test Shapely 2.0 (currently 2.0b1) alongside GeoPandas
+
+Note that if you also have PyGEOS installed, you need to set an environment variable
+(`USE_PYGEOS=0`) before importing geopandas to actually test Shapely 2.0 features instead of PyGEOS. See
+https://geopandas.org/en/latest/getting_started/install.html#using-the-optional-pygeos-dependency
+for more details.
+
 New features and improvements:
 
 - Added ``normalize()`` method from shapely to GeoSeries/GeoDataframe (#2537)
@@ -12,10 +23,16 @@ New features and improvements:
   'color', 'legend', 'categorical', 'scheme', 'k', 'vmin' and 'vmax' arguments (#2590)
 
 Deprecations and compatibility notes:
+- Accessing the `crs` of a `GeoDataFrame` without active geometry column was deprecated and this now raises an AttributeError (#2578).
+
+- resolve ``matplotlib.cm`` warning in ``.explore()`` (#2596)
 
 Bug fixes:
-
-Notes on (optional) dependencies:
+- Fix cryptic error message in ``geopandas.clip()`` when clipping with an empty geometry (#2589)
+- Accessing `gdf.geometry` where the active geometry column is missing, and a column named `"geometry"` is present 
+  will now raise an `AttributeError`, rather than returning `gdf["geometry"]` (#2575) 
+- Combining GeoSeries/GeoDataFrames with ``pandas.concat`` will no longer silently
+  override CRS information if not all inputs have the same CRS (#2056).
 
 Version 0.11.1 (July 24, 2022)
 ------------------------------